requirements.txt
configparser
mysql-connector-python
flask_apscheduler
flask
wheel
# rpi-gpio
hx711
python-dotenv